Added Hackers section and link with register page
